requirements.txt
pillow~=8.4.0
opencv-python
numpy